services/ai-engine/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.0
pydantic-settings==2.2.1
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.4
aiokafka==0.10.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
prophet==1.1.5
tensorflow==2.16.1
joblib==1.4.2
httpx==0.27.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
structlog==24.1.0
prometheus-fastapi-instrumentator==7.0.0
pytest==8.1.2
pytest-asyncio==0.23.6
aiofiles==23.2.1
